tests/requirements.txt
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-html==4.1.1
pytest-xdist==3.5.0
httpx==0.25.2
psycopg2-binary==2.9.9
redis==5.0.1
docker==6.1.3
testcontainers==3.7.1
faker==20.1.0
aiofiles==23.2.1
python-dotenv==1.0.0